air-quality-producer/requirements.txt
kafka-python
geopy
requests
numpy
faker